updated creation of gdms_markers_onmap table, changed data types from float to double of two fields
git-svn-id: https://efficio.atlassian.net/svn/GCP/trunk/IBDBScripts@2402 4ebb2e00-0138-4608-bc6e-edf3948d4863

+++ IBDBv1_GDMS.sql
@@ -268,8 +268,8 @@ DROP TABLE IF EXISTS `gdms_markers_onmap`;
 CREATE TABLE `gdms_markers_onmap` (
   `map_id` int(11) NOT NULL DEFAULT '0',
   `marker_id` int(11) NOT NULL DEFAULT '0',
-  `start_position` float DEFAULT NULL,
-  `end_position` float DEFAULT NULL,
+  `start_position` DOUBLE DEFAULT NULL,
+  `end_position` DOUBLE DEFAULT NULL,
   `map_unit` char(4) DEFAULT NULL,
   `linkage_group` varchar(50) DEFAULT NULL,
   KEY `fk_markerlm_markerid` (`marker_id`),
